requirements.txt
torch
torchvision
ultralytics
onnxruntime-gpu
pandas
PyYAML
opencv-python
Pillow
tqdm
numpy
